+++ WebSocket.js
@@ -4,22 +4,22 @@ function empty() {}
 
 export class websocket {
 	constructor() {
-		this.on_message = empty;
-		this.on_open = empty;
-		this.on_close = empty;
-		this.uri = null;
+		this._on_message = empty;
+		this._on_open = empty;
+		this._on_close = empty;
+		this._uri = null;
 	}
 
-	uri(uri) { this.uri = uri; return this; }
-	on_message(on_message) { this.on_message = on_message; return this; }
-	on_open(on_open) { this.on_open = on_open; return this; }
-	on_close(on_close) { this.on_close = on_close; return this; }
+	uri(uri) { this._uri = uri; return this; }
+	on_message(on_message) { this._on_message = on_message; return this; }
+	on_open(on_open) { this._on_open = on_open; return this; }
+	on_close(on_close) { this._on_close = on_close; return this; }
 
 	connect() {
-		let ws = new WebSocket('ws://' + this.uri);
-		ws.onopen = this.on_open;
-		ws.onmessage = this.on_message;
-		ws.onclose = this.on_close;
+		let ws = new WebSocket('ws://' + this._uri);
+		ws.onopen = this._on_open;
+		ws.onmessage = this._on_message;
+		ws.onclose = this._on_close;
 		return ws;
 	}
 }